requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.3
numpy==1.25.2
yfinance==0.2.28
openai==1.3.5
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
plotly==5.17.0
selenium==4.15.2
webdriver-manager==4.0.1
lxml==4.9.3
python-dateutil==2.8.2
ta==0.10.2
newsapi-python==0.2.7 